Organize util test directories
